requirements/dev.txt
py-cord[voice]
python-dotenv
google-cloud-texttospeech
azure-cognitiveservices-speech
httpx
cloudscraper
redis
flask
gunicorn
metadata_parser
beautifulsoup4
lxml
psycopg2-binary
black
natsort
